Refactor the backend. Closes #145
