+++ lib/api/email.js
@@ -1,9 +1,14 @@
 const routes = require('express').Router();
 const {build, send} = require('revolutionuc-emails');
 const {readOnlyAPIKeyValidator, apiKeyValidator, eventDeadlineValidator} = require('./mwValidators');
+const {Registrant} = require('../models/registrant');
+const crypto = require ('crypto');
 
 const mailgunApiKey = process.env.MAILGUN_API_KEY;
 const mailgunDomain = process.env.MAILGUN_DOMAIN;
+
+const userCryptoAlgorithm = 'aes-256-ctr';
+const userCryptoPassword = process.env.MONGO_SALT;
 
 /**
  * Email route handlers
@@ -25,7 +30,7 @@ routes.get('/sendConfirmation', apiKeyValidator, eventDeadlineValidator, (reques
   Registrant.find({emailVerified: true, waitList: false, receivedAttendanceConfirmationEmail: false}, (err, registrants) => {
     if (err) {
       console.error(err);
-      response.status(400).send('Sorry, something went wrong');
+      return response.status(400).send('Sorry, something went wrong');
     }
 
     sendConfirmationEmails(request, registrants);
@@ -51,7 +56,7 @@ routes.get('/sendConfirmationToSittingDucks', apiKeyValidator, eventDeadlineVali
   Registrant.find({receivedAttendanceConfirmationEmail: true, attendanceConfirmed: 'NA'}, (err, registrants) => {
     if (err) {
       console.error(err);
-      response.status(400).send('Sorry, something went wrong');
+      return response.status(400).send('Sorry, something went wrong');
     }
 
     sendConfirmationEmails(request, registrants);
@@ -76,7 +81,7 @@ routes.get('/sendConfirmationToWaitlisted', apiKeyValidator, eventDeadlineValida
   Registrant.find({emailVerified: true, waitList: true}, (err, registrants) => {
     if (err) {
       console.error(err);
-      response.status(400).send('Sorry, something went wrong');
+      return response.status(400).send('Sorry, something went wrong');
     }
 
     Registrant.update({emailVerified: true, waitList: true}, {waitList: false}, {multi: true}, (err, numberAffected, rawResponse) => {
@@ -110,7 +115,7 @@ routes.get('/sendConfirmationToIndividual', apiKeyValidator, (request, response)
     Registrant.update({email: userEmail}, {waitList: false}, (err, numberAffected, rawResponse) => {
       if (err) {
         console.error(err);
-        response.status(400).send('Sorry, something went wrong');
+        return response.status(400).send('Sorry, something went wrong');
       }
 
       sendConfirmationEmails(request, registrants);
@@ -133,7 +138,7 @@ routes.get('/sendInfoEmail1', apiKeyValidator, eventDeadlineValidator, (request,
   Registrant.find({emailVerified: true}, (err, registrants) => {
     if (err) {
       console.error(err);
-      response.status(400).send('Sorry, something went wrong');
+      return response.status(400).send('Sorry, something went wrong');
     }
 
     sendInfoEmail1(request, registrants);
