+++ lib/api/index.js
@@ -1,4 +1,10 @@
 const routes = require('express').Router();
+const mongoose = require('mongoose');
+
+mongoose.connect(process.env.MONGO_URL);
+mongoose.connection.on('error', console.error.bind(console, 'connection error:'));
+mongoose.connection.once('open', _ => {console.log('Connected to mongoose in registration controller')});
+
 const registrationRoutes = require('./registration');
 const emailRoutes = require('./email');
