+++ lib/api/registration.js
@@ -2,14 +2,13 @@ const AWS = require('aws-sdk');
 const s3 = new AWS.S3();
 const crypto = require('crypto');
 const multer = require('multer');
-const mongoose = require('mongoose');
 const routes = require('express').Router();
 const {build, send} = require('revolutionuc-emails');
 const {readOnlyAPIKeyValidator, apiKeyValidator, eventDeadlineValidator} = require('./mwValidators');
+const {Registrant} = require('../models/registrant');
 
 const mailgunApiKey = process.env.MAILGUN_API_KEY;
 const mailgunDomain = process.env.MAILGUN_DOMAIN;
-const MONGO_URL = process.env.MONGO_URL;
 const WAITLIST_THRESHOLD = process.env.WAITLIST_THRESHOLD;
 
 const storage = multer.diskStorage({
@@ -23,119 +22,8 @@ const storage = multer.diskStorage({
 
 const upload = multer({storage, limits: { fileSize: 20000000, files: 1 }}); // 20mb max
 
-mongoose.connect(MONGO_URL);
-
-mongoose.connection.on('error', console.error.bind(console, 'connection error:'));
-mongoose.connection.once('open', _ => {console.log('Connected to mongoose in registration controller')});
-
 const userCryptoAlgorithm = 'aes-256-ctr';
 const userCryptoPassword = process.env.MONGO_SALT;
-
-const registrantScheme = mongoose.Schema({
-  firstName: {
-    type: String,
-    required: true
-  },
-  lastName: {
-    type: String,
-    required: true
-  },
-  email: {
-    type: String,
-    required: true
-  },
-  phoneNumber: {
-    type: String,
-    required: true
-  },
-  school: {
-    type: String,
-    required: true
-  },
-  schoolOther: {
-    type: String,
-    default: "off"
-  },
-  major: {
-    type: String,
-    required: true
-  },
-  education: {
-    type: String,
-    required: true
-  },
-  gender: {
-    type: String,
-    required: true
-  },
-  ethnicity: {
-    type: String,
-    required: true
-  },
-  howYouHeard: {
-    type: String
-  },
-  hackathons: {
-    type: String
-  },
-  shirtSize: {
-    type: String,
-    required: true
-  },
-  githubUsername: {
-    type: String
-  },
-  dob: {
-    type: String,
-    required: true
-  },
-  registrationTime: {
-    type: String,
-    required: true
-  },
-  waitList: {
-    type: Boolean,
-    default: false
-  },
-  checkedIn: {
-    type: Boolean,
-    default: false
-  },
-  emailVerified: {
-    type: Boolean,
-    default: false
-  },
-  attendanceConfirmed: {
-    type: String,
-    default: "NA",
-  },
-  receivedAttendanceConfirmationEmail: {
-    type: Boolean,
-    default: false
-  },
-  vegetarian: {
-    type: Boolean,
-    default: false
-  },
-  vegan: {
-    type: Boolean,
-    default: false
-  },
-  peanutAllergy: {
-    type: Boolean,
-    default: false
-  },
-  glutenFree: {
-    type: Boolean,
-    default: false
-  },
-  otherDietaryRestrictions: {
-    type: String,
-    default: ''
-  }
-});
-
-const Registrant = mongoose.model('Registrant', registrantScheme);
 
 /**
  * Registration route handlers
@@ -258,7 +146,11 @@ routes.get('/verify', eventDeadlineValidator, (request, response) => {
   const userEmail = dec;
 
   Registrant.update({email: userEmail}, {emailVerified: true}, (err, numberAffected, rawResponse) => {
-    if (err) response.status(400).send('Sorry, something went wrong');
+    if (err) {
+      console.log(err);
+      return response.status(400).send('Sorry, something went wrong');
+    }
+
     response.status(200).redirect('https://revolutionuc.com/email-verified');
   });
 });
@@ -273,7 +165,6 @@ routes.get('/confirm', eventDeadlineValidator, (request, response) => {
   dec += decipher.final('utf8');
 
   const userEmail = dec;
-
   const confirmValue = request.query.confirm;
 
   const emailData = {subject: 'Welcome to RevolutionUC!', shortDescription: 'What you need to know for RevolutionUC'};
@@ -282,15 +173,21 @@ routes.get('/confirm', eventDeadlineValidator, (request, response) => {
     return response.status(400).send('You must confirm yes or no');
   }
 
-  Registrant.findOne({email: userEmail}, (err, registrant) => {
-    if (err) response.status(400).send('Sorry, something went wrong');
+  Registrant.findOne({email: userEmail, emailVerified: true}, (err, registrant) => {
+    if (err || !registrant) {
+      console.error(err);
+      return response.status(400).send('Sorry, something went wrong');
+    }
 
     if (registrant && (registrant.attendanceConfirmed == 'YES' || registrant.attendanceConfirmed == 'NO')) {
       return response.send('You have already confirmed your attendance');
     }
 
     Registrant.update({email: userEmail}, {attendanceConfirmed: confirmValue}, (err, numberAffected, rawResponse) => {
-      if (err) response.status(400).send('Sorry, something went wrong');
+      if (err) {
+        console.error(err);
+        return response.status(400).send('Sorry, something went wrong');
+      }
 
       if (confirmValue == 'YES') {
         emailData.firstName = registrant.firstName;
@@ -298,14 +195,10 @@ routes.get('/confirm', eventDeadlineValidator, (request, response) => {
           .then(html => {
             return send(mailgunApiKey, mailgunDomain, 'RevolutionUC <info@revolutionuc.com>', registrant.email, emailData.subject, html);
           })
-          .then(() => {
-            if (err) return console.error(err);
-            return response.status(201).json(savedRegistrant);
-          }, e => {
+          .catch(e => {
             console.log(e);
             return response.status(400).send(e);
           }); // build().then()
-
       }
 
       // Send them to the proper page after they confirm or deny attendence
@@ -325,7 +218,10 @@ routes.get('/checkIn', apiKeyValidator, (request, response) => {
   const userEmail = request.query.userEmail;
 
   Registrant.update({email: userEmail}, {checkedIn: true}, (err, numberAffected, rawResponse) => {
-    if (err) return response.status(400).send('Sorry, something went wrong');
+    if (err) {
+      console.error(err);
+      return response.status(400).send('Sorry, something went wrong');
+    }
 
     response.json({checkedIn: true});
   });
@@ -333,7 +229,11 @@ routes.get('/checkIn', apiKeyValidator, (request, response) => {
 
 routes.get('/all', readOnlyAPIKeyValidator, (request, response) => {
   Registrant.find((err, allRegistrants) => {
-    if (err) return console.error(err);
+    if (err) {
+      console.error(err);
+      return response.status(400).send('Sorry, something went wrong');
+    }
+
     response.json(allRegistrants);
   });
 });
